int to float

+++ js/setup_draw.js
@@ -49,8 +49,8 @@ function setup() {
 
 //LOOP
 function draw() {
-    k = parseInt(sliderkill.value);
-    f = parseInt(sliderfeed.value);
+    k = parseFloat(sliderkill.value);
+    f = parseFloat(sliderfeed.value);
     for (let x = 1; x < width - 1; x++) {
         for (let y = 1; y < height - 1; y++) {
             let a = grid[x][y].a
